Refactored CMakeLists's to improve isolation.
Add BUILD_SERVER option.
Update environment.yml.

+++ polymetis/src/clients/empty_statistics_client.cpp
@@ -131,7 +131,7 @@ private:
     ClientContext context;
     Status status = stub_->ControlUpdate(&context, state_, &torque_command);
     if (!status.ok()) {
-      spdlog::error("SendCommand failed: {}", status.error_code());
+      spdlog::error("SendCommand failed: {}", static_cast<int>(status.error_code()));
       return false;
     }
     return true;
